requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==1.11.26
edx-opaque-keys==2.0.0
funcsigs==1.0.2
idna==2.8
mock==3.0.5
pbr==5.4.4
pkg-resources==0.0.0
pycodestyle==2.5.0
pymongo==3.9.0
pytz==2019.3
requests==2.22.0
six==1.13.0
stevedore==1.31.0
Unidecode==1.1.1
urllib3==1.25.6
